gonna merge

+++ src/sig_handler.c
@@ -10,7 +10,18 @@
 /* SIGNAL HEADERS */
 #include <signal.h>
 
-extern int globalProcess;
+int globalProcess = 0;
+int check_process = 0;
+
+void setGlobalProcess(int pgid){
+    globalProcess = pgid;
+    check_process = 1;
+}
+
+void resetGlobalProcess(void){
+    globalProcess = 0;
+    check_process = 0;
+}
 
 //Handler para quando o processo pai recebe SIGINT
 void sigint_handler(int signo){
@@ -35,7 +46,7 @@ void sigint_handler(int signo){
   //dup2(saved_stdout, 1);
   //close(saved_stdout);
 
-  killpg(globalProcess,SIGSTOP);  //Envia a ele mesmo o sinal, porque é uncatchable
+  if(check_process) killpg(globalProcess,SIGSTOP);  //Envia a ele mesmo o sinal, porque é uncatchable
 
   //killpg(pgid,SIGCONT);
   //Deve enviar SIGSTOP para todos os processos filho
@@ -45,9 +56,10 @@ void sigint_handler(int signo){
 
   int n = read(STDIN_FILENO,ch,256);
   if((n == 2 ) && (ch[0] == 'y')){
-      killpg(0,SIGTERM); // Termina todos os processos parados anteriormente
+      if(check_process) killpg(globalProcess,SIGTERM); // Termina todos os processos parados anteriormente
+      raise(SIGTERM);
   }
   else{
-      killpg(globalProcess,SIGCONT); // Continua todos os processos parados anteriormente
+      if(check_process) killpg(globalProcess,SIGCONT); // Continua todos os processos parados anteriormente
   }
 }
